[IMP][FIX][CLIENTE-2744] Corrección de método post_init_hook
